+++ src/main/java/com/example/nuggetbe/entity/Connection.java
@@ -4,11 +4,13 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
+@Entity
 @AllArgsConstructor
 @NoArgsConstructor
+@Setter
 @Getter
-@Entity
 public class Connection {
 
     @Id
