+++ src/main/java/com/example/nuggetbe/repository/ConnectionRepository.java
@@ -1,0 +1,8 @@
+package com.example.nuggetbe.repository;
+import com.example.nuggetbe.entity.Connection;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ConnectionRepository extends JpaRepository<Connection, Long> {
+}
